+++ src/test/java/com/epam/reportportal/junit/testcaseid/TestCaseIdStaticStandardParametersTest.java
@@ -16,7 +16,7 @@
 
 package com.epam.reportportal.junit.testcaseid;
 
-import com.epam.reportportal.junit.ParallelRunningHandler;
+import com.epam.reportportal.junit.ReportPortalListener;
 import com.epam.reportportal.junit.features.parameters.StandardParametersSimpleTest;
 import com.epam.reportportal.junit.utils.TestUtils;
 import com.epam.reportportal.service.ReportPortal;
@@ -51,7 +51,7 @@ public class TestCaseIdStaticStandardParametersTest {
 		client = mock(ReportPortalClient.class);
 		TestUtils.mockLaunch(client, launchId, suiteId, classId, methodId);
 		TestUtils.mockLogging(client);
-		ParallelRunningHandler.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
+		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}
 
 	@Test